requirements.txt
# Dependencies for Veo3 Video Generator Application

# PyQt6 - GUI Framework
PyQt6>=6.7.0

# Google AI SDK (optional - app works in mock mode without it)
google-genai>=1.48.0

# Environment variables
python-dotenv>=1.0.1

# Async HTTP Client
aiohttp>=3.9.5

# Image processing
Pillow>=10.3.0

# Additional libraries
requests>=2.31.0

# Development tools (optional, has version conflicts with PyQt6 6.7+)
# PyQt6-tools==6.4.2.3.3
